HCF or GCD
In the following tutorial, we will understand how to find Highest Common Factor (HCF) in the Python programming language.

But before we get started, let us briefly discuss about HCF.

HCF: Highest Common Factor
Highest Common Factor or Greatest Common Divisor of two or more integers when at least one of them is not zero is the largest positive integer that evenly divides the numbers without a remainder. For example, the GCD of 8 and 12 is 4.

For example:

We have two integers 8 and 12. Let's find the HCF.

The divisors of 8 are:

1, 2, 4, 8  
The divisors of 12 are:

1, 2, 3, 4, 6, 12  
HCF /GCD is the greatest common divisor. So HCF of 8 and 12 are 4.

Now, let us consider an example based on finding HCF of two given numbers.